requirements.txt
fastapi[all]
asyncpg
python-dotenv